% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arreglar_df_Eurostat.R
\name{arreglar_df_Eurostat}
\alias{arreglar_df_Eurostat}
\title{arregla una base de datos sacada de Eurostat}
\usage{
arreglar_df_Eurostat(df, df_l)
}
\arguments{
\item{df}{Un dataframe}
}
\value{
Un nuevo df
}
\description{
te retorna un df arreglado (con los codigos y labels)
}
\examples{
arreglar_df_Eurostat(df)
}

