% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimales_df_pjp.R
\name{decimales_df_pjp}
\alias{decimales_df_pjp}
\title{deja los decimales q le digas a un df}
\usage{
decimales_df_pjp(df, nn = 2)
}
\arguments{
\item{df}{Un dataframe}

\item{nn}{el nº de decimales (Default = 2)}
}
\value{
Un nuevo \code{df} con sus valores redondeados a \code{nn}
}
\description{
decimales_df_pjp() te retorna un df con los decimales que le digas
decimales_df_pjp function redondea los valores de un dataframe. Le metes un df y redondea sus valores
}
\examples{
decimales_df_pjp(cars, nn = 2)
}

