% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_nice_pjp.R
\name{round_nice_pjp}
\alias{round_nice_pjp}
\title{round numbers nicely}
\usage{
round_nice_pjp(..., nn = 2)
}
\arguments{
\item{nn}{(digitos, default=2)}
}
\description{
Ya lo he dicho round numbers nicely
This function rounds data in a nice way. See: ttps://twitter.com/hspter/status/314858331598626816
}
\examples{
 round_nice_pjp(c(-0.0002, 0.0002, 0.5, -0.5, -0.002), dig=3))   this should return c("0.000", "0.000", "0.500", "-0.500", "-0.002")
}

