% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_unicos_df_pjp.R
\name{val_unicos_df_pjp}
\alias{val_unicos_df_pjp}
\title{unique values in each colum}
\usage{
val_unicos_df_pjp(df)
}
\arguments{
\item{df}{(a data frame)}
}
\description{
val_unicos_df_pjp() te retorna un df con los valores(unicos) que existen en cada columna de un df
}
\examples{
val_unicos_df_pjp(cars)
}

